v5.x-remove-runtimecosting: removed SigmaDslOverArrays.scala
